// src/api/generated/models/pizzaDetailsDto.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * PizzaApp | v1
 * OpenAPI spec version: 1.0.0
 */
import type { IngredientDto } from './ingredientDto';
import type { LookUpItemDto } from './lookUpItemDto';

export interface PizzaDetailsDto {
  id?: string;
  menuId?: string;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  price?: number | string;
  /** @nullable */
  imageUrl?: string | null;
  style: LookUpItemDto;
  dough: LookUpItemDto;
  baseSauce: LookUpItemDto;
  thickness: LookUpItemDto;
  shape: LookUpItemDto;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$ */
  weightGrams?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$ */
  kcal?: number | string;
  /**
   * @nullable
   * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
   */
  diameterCm?: number | string | null;
  /**
   * @nullable
   * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
   */
  widthCm?: number | string | null;
  /**
   * @nullable
   * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$
   */
  lengthCm?: number | string | null;
  ingredients?: IngredientDto[];
}